Use Brewfiles for more efficient installation.
Move most formulae and casks into two Brewfiles:

- Brewfile-brews
- Brewfile-casks

This speeds up installation somewhat (not having the overhead of calling brew for each install), but is much more noticeable when re-running these install scripts with only minor changes, when up-to-date packages take almost no time at all.  It's not as colorful or animated, but much faster.